+++ src/test/java/org/kamillion/hateoflux/linkbuilder/SpringControllerLinkBuilderTest.java
@@ -20,6 +20,33 @@ class SpringControllerLinkBuilderTest {
 
     @ParameterizedTest
     @NullAndEmptySource
+    void givenPostMappingWithCompositeCollectionAsQueryParameter_whenProvidedEmptyOrNullValues_thenLinkUnchanged(List<String> args) {
+        //GIVEN & WHEN
+        Link link = linkTo(DummyController.class, c -> c.postMappingWithCollectionAsQueryParameter(args));
+
+        //THEN
+        assertThat(link.getHref()).isEqualTo("/dummy/names");
+    }
+
+    @ParameterizedTest
+    @CsvSource(delimiter = ';', value = { //
+            "val1; ?names=val1",
+            "val1|val2; ?names=val1&names=val2"
+    })
+    void givenPostMappingWithCompositeCollectionAsQueryParameter_whenProvidedValues_thenLinkIsCorrect(
+            String args, String expectedQueryParameters) {
+        //GIVEN & WHEN
+        List<String> argsAsList = Arrays.stream(args.split("\\|")).toList();
+        Link link = linkTo(DummyController.class,
+                c -> c.postMappingWithCompositeCollectionAsQueryParameter(argsAsList));
+
+        //THEN
+        assertThat(link.getHref()).isEqualTo("/dummy/names" + expectedQueryParameters);
+    }
+
+
+    @ParameterizedTest
+    @NullAndEmptySource
     void givenPostMappingWithCollectionAsQueryParameter_whenProvidedEmptyOrNullValues_thenLinkUnchanged(List<String> args) {
         //GIVEN & WHEN
         Link link = linkTo(DummyController.class, c -> c.postMappingWithCollectionAsQueryParameter(args));